가속도와 가속도 가중치 코드를 분리
