+++ AR_illusion_training/Assets/scripts/VelocityText.cs
@@ -1,0 +1,51 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using TMPro;
+using Unity.VisualScripting;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+using UnityEngine.UI;
+using System.Runtime.InteropServices;
+
+public class VelocityText : MonoBehaviour
+{
+#if UNITY_IOS
+    [DllImport("__Internal")]
+    private static extern void _startTracking(string objectName);
+#endif
+    public TextMeshProUGUI OutputText;
+    bool isTracking = false;
+
+    
+    void Start()
+    {
+        StartSpeedTracking();
+    }
+
+    void Update()
+    {
+        
+    }
+
+
+    public void StartSpeedTracking()
+    {
+        if (this.isTracking == false)
+        {
+            this.isTracking = true;
+#if UNITY_IOS
+            Debug.Log($"GameObjectName: {gameObject.name}");
+            _startTracking(gameObject.name);
+#else
+            Debug.Log("No iOS Device Found");
+#endif
+        }
+
+    }
+
+    public void SpeedCallBackMethod(string speed)
+    {
+        Debug.Log($"유니티에서 받은 스피드: {speed}");
+    }
+}
