+++ AR_illusion_training/Assets/scripts/VelocityText.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e691e61b45baa164f8b8f7bde52d7f83
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
