+++ AR_illusion_training/Assets/scripts/VelocityWeightText.cs
@@ -10,18 +10,12 @@ using System.Runtime.InteropServices;
 
 public class VelocityWeightText : MonoBehaviour
 {
-#if UNITY_IOS
-    [DllImport("__Internal")]
-    private static extern void _startTracking(string objectName);
-#endif
-
     public TextMeshProUGUI OutputText;
     public Scrollbar VelocityWeightScrollbar;
     float VelocityWeight = 1f;
-    bool isTracking = false;
-
+    
     void Start(){
-        StartSpeedTracking();
+        
     }
     void Update(){
         UpdateVelocity();
@@ -33,23 +27,5 @@ public class VelocityWeightText : MonoBehaviour
         OutputText.text = VelocityWeight.ToString(); // Velocity 값을 문자열로 변환하여 표시
     }
 
-    public void StartSpeedTracking()
-    {
-        if (this.isTracking == false)
-        {
-            this.isTracking = true;
-#if UNITY_IOS
-            Debug.Log($"GameObjectName: {gameObject.name}");
-            _startTracking(gameObject.name);
-#else
-            Debug.Log("No iOS Device Found");
-#endif
-        }
-
-    }
-
-    public void SpeedCallBackMethod(string speed)
-    {
-        Debug.Log($"유니티에서 받은 스피드: {speed}");
-    }
+   
 }
